% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.gee1step}
\alias{coef.gee1step}
\title{Coef method for gee1step model fits}
\usage{
\method{coef}{gee1step}(object, ...)
}
\arguments{
\item{object}{a fitted model object of class "gee1step".}

\item{...}{currently disregarded}
}
\value{
a vector of length \eqn{p} that represents the parameter estimates.
}
\description{
Coef method for gee1step model fits
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData_binomial,
  cluster = "site", family="binomial")
coef(geefit)

}
