% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.gee1step}
\alias{confint.gee1step}
\title{Confidence intervals for gee1step model parameters}
\usage{
\method{confint}{gee1step}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{a fitted gee1step model object}

\item{parm}{a specification of which parameters are to be given confidence
intervals, either a vector of numbers or a vector of names. If missing, all
parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{currently disregarded}
}
\value{
nothing is returned
}
\description{
Confidence intervals for gee1step model parameters
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
confint(geefit)

}
