% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee1step.R
\name{gee1step}
\alias{gee1step}
\title{Estimate parameters using one-step algorithm}
\usage{
gee1step(formula, data, cluster)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{a required data frame or data.table containing the variables in the model.}

\item{cluster}{the name of the field that identifies the clusters.}
}
\value{
a "gee1step" object
}
\description{
Estimate parameters using one-step algorithm
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
geefit

}
