% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.gee1step}
\alias{predict.gee1step}
\title{Predict method for gee1step model fits}
\usage{
\method{predict}{gee1step}(object, newdata = NULL, type = "link", ...)
}
\arguments{
\item{object}{a fitted model object of class "gee1step".}

\item{newdata}{an optional data set with new covariate data. If no new data are
provided, predictions are provided for the original data set.}

\item{type}{the type of prediction required. The default ("link") is on the scale of
the linear predictors; the alternative "response" is on the scale of the
response variable. For this model (a binomial model) the default predictions
are of log-odds (probabilities on logit scale) and type = "response" gives
the predicted probabilities.}

\item{...}{currently disregarded}
}
\value{
a vector of predictions
}
\description{
Predict method for gee1step model fits
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData_binomial,
  cluster = "site", family="binomial")
predict(geefit, sampData_binomial)

}
