% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.gee1step}
\alias{print.summary.gee1step}
\title{print method for summary.gee1step}
\usage{
\method{print}{summary.gee1step}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{a summary.gee1step object.}

\item{digits}{the number of significant digits.}

\item{...}{currently disregarded}
}
\value{
nothing is returned
}
\description{
print method for summary.gee1step
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
print(geefit)

}
