% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampData_binomial}
\alias{sampData_binomial}
\title{Simulated data set}
\format{
A data.table with 9788 rows and 6 variables:
\describe{
  \item{id}{individual record identifier}
  \item{site}{site indentifer}
  \item{x1}{continuous covariate}
  \item{x2}{continuous covariate}
  \item{x3}{continuous covariate}
  \item{y}{binary outcome}
}
}
\usage{
sampData_binomial
}
\description{
A simulated dataset containing a binary outcome data clustered at the site level and
three covariates.
}
\keyword{datasets}
