% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.gee1step}
\alias{summary.gee1step}
\title{Summarizing gee1step model fits}
\usage{
\method{summary}{gee1step}(object, ...)
}
\arguments{
\item{object}{a fitted model object of class "gee1step".}

\item{...}{currently disregarded}
}
\value{
a \eqn{p \times 4} matrix with columns for the estimated coefficient,
its standard error, z-statistic and corresponding (two-sided) p-value, the
ICC, and summary of cluster sizes.
}
\description{
Summarizing gee1step model fits
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
summary(geefit)

}
