% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.gee1step}
\alias{vcov.gee1step}
\title{vcov method for gee1step model fits}
\usage{
\method{vcov}{gee1step}(object, ...)
}
\arguments{
\item{object}{a fitted model object of class "gee1step".}

\item{...}{currently disregarded}
}
\value{
a matrix of dimension \eqn{p \timts p} that represents the parameter
variance-covariance matrix
}
\description{
vcov method for gee1step model fits
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
vcov(geefit)

}
