% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee1step.R
\name{predict.gee1step}
\alias{predict.gee1step}
\title{Estimate parameters using one-step algorithm}
\usage{
\method{predict}{gee1step}(object, data, type = "link", ...)
}
\arguments{
\item{object}{a fitted model object of class "gee1step".}

\item{data}{a required data frame or data.table containing the variables
in the model.}

\item{type}{the type of prediction required. The default ("link") is on the scale of
the linear predictors; the alternative "response" is on the scale of the
response variable. For this model (a binomial model) the default predictions
are of log-odds (probabilities on logit scale) and type = "response" gives
the predicted probabilities.}

\item{...}{currently disregarded}
}
\value{
a vector of predictions
}
\description{
Estimate parameters using one-step algorithm
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
predict(geefit, sampData)

}
