% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.gee1step}
\alias{print.gee1step}
\title{print method for gee1step model fits}
\usage{
\method{print}{gee1step}(x, ...)
}
\arguments{
\item{x}{a fitted model object of class "gee1step".}

\item{...}{currently disregarded}
}
\value{
nothing is returned
}
\description{
print method for gee1step model fits
}
\examples{
geefit <- gee1step(y ~ x1 + x2 + x3, data = sampData, cluster = "site")
print(geefit)

}
